pyproject.toml
[tool.poetry]
name = "examen-ksp"
version = "0.1.0"
description = ""
authors = ["alexhdz1 <alex.hdz.c@ciencias.unam.mx>"]

[tool.poetry.dependencies]
python = "^3.9"
django-admin = "^2.0.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
Django = "^4.1.4"
mysqlclient = "^2.1.1"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.21.0"
decouple = "^0.0.7"
python-decouple = "^3.6"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
